pushing unrelated functions into utils
